% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_functions.R
\name{calculate_exposure_index}
\alias{calculate_exposure_index}
\title{Calculer l'indice d'exposition climatique}
\usage{
calculate_exposure_index(
  heatwave_df,
  drought_df,
  frost_df,
  w_heat = 0.4,
  w_drought = 0.4,
  w_frost = 0.2,
  agri_importance = 1
)
}
\arguments{
\item{heatwave_df}{Dataframe résultat de \code{\link{calculate_heatwave_index}}.}

\item{drought_df}{Dataframe résultat de \code{\link{calculate_drought_index}}.}

\item{frost_df}{Dataframe résultat de \code{\link{calculate_frost_risk}}.}

\item{w_heat}{Poids de l'indice chaleur (entre 0 et 1). Par défaut : 0.4.}

\item{w_drought}{Poids de l'indice sécheresse. Par défaut : 0.4.}

\item{w_frost}{Poids de l'indice gel. Par défaut : 0.2.}

\item{agri_importance}{Importance agricole de la zone (0-1). Par défaut : 1.}
}
\value{
Un dataframe avec les colonnes :
\describe{
\item{year}{Année}
\item{heat_score}{Score normalisé chaleur (0-1)}
\item{drought_score}{Score normalisé sécheresse (0-1)}
\item{frost_score}{Score normalisé gel (0-1)}
\item{exposure_index}{Indice composite (0-1)}
\item{exposure_class}{Classe : "faible", "moyen", "élevé"}
}
}
\description{
Calcule un indice composite d'exposition climatique en combinant les scores
de chaleur, de sécheresse et de gel, pondérés par l'importance agricole.
Produit trois classes d'exposition : faible, moyen, élevé.
}
\examples{
data(climate_morocco)
hw <- calculate_heatwave_index(climate_morocco, threshold = 38)
dr <- calculate_drought_index(climate_morocco)
fr <- calculate_frost_risk(climate_morocco)
expo <- calculate_exposure_index(hw, dr, fr)
head(expo)

}
