% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_reporting.R
\name{generate_risk_atlas}
\alias{generate_risk_atlas}
\title{Generer un atlas climatique automatique}
\usage{
generate_risk_atlas(
  climate_data,
  exposure_df,
  output_file = "climate_atlas.html",
  title = "Atlas des Risques Climatiques Agricoles",
  region = "Zone d'etude",
  open_browser = FALSE
)
}
\arguments{
\item{climate_data}{Dataframe climatique de base.}

\item{exposure_df}{Dataframe d'indice d'exposition (resultat de
\code{\link{calculate_exposure_index}}).}

\item{output_file}{Nom du fichier HTML de sortie. Par defaut :
\code{"climate_atlas.html"}.}

\item{title}{Titre du rapport.}

\item{region}{Nom de la region etudiee.}

\item{open_browser}{Ouvrir le rapport dans le navigateur apres generation.}
}
\value{
Chemin du fichier HTML genere (invisible).
}
\description{
Genere un rapport HTML complet comprenant les indices climatiques,
les tendances, les cartes de risque et les recommandations d'adaptation.
}
\examples{
\dontrun{
data(climate_morocco)
hw   <- calculate_heatwave_index(climate_morocco)
dr   <- calculate_drought_index(climate_morocco)
fr   <- calculate_frost_risk(climate_morocco)
expo <- calculate_exposure_index(hw, dr, fr)
generate_risk_atlas(climate_morocco, expo,
                    output_file = "atlas_maroc.html",
                    region = "Maroc")
}

}
