% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{summarize_extremes}
\alias{summarize_extremes}
\title{Résumer les extrêmes climatiques}
\usage{
summarize_extremes(
  heatwave_df = NULL,
  drought_df = NULL,
  frost_df = NULL,
  ref_period = NULL
)
}
\arguments{
\item{heatwave_df}{Dataframe résultat de \code{\link{calculate_heatwave_index}}.}

\item{drought_df}{Dataframe résultat de \code{\link{calculate_drought_index}}.}

\item{frost_df}{Dataframe résultat de \code{\link{calculate_frost_risk}}.}

\item{ref_period}{Vecteur de deux années pour la période de référence.
Ex : \code{c(1981, 2010)}.}
}
\value{
Un dataframe de synthèse avec des statistiques clés par indice.
}
\description{
Produit un tableau de synthèse statistique des indices d'extrêmes
climatiques : fréquence, intensité, durée et anomalies.
}
\examples{
data(climate_morocco)
hw <- calculate_heatwave_index(climate_morocco, threshold = 38)
dr <- calculate_drought_index(climate_morocco)
fr <- calculate_frost_risk(climate_morocco)
summary_df <- summarize_extremes(hw, dr, fr)
print(summary_df)

}
